// sourcecode/BTrak.Packages/snova-dashboard-module/Dashboard-module-components/dashboard-components/src/lib/snovasys-dashboard-management/models/historicalWorkReport.ts
import { SearchCriteriaInputModelBase } from './search-criteria-input-base.model';

export class HistoricalWorkReportModel extends SearchCriteriaInputModelBase {
    userId: string;
    dateFrom: Date;
    dateTo: Date;
    createDateTo: Date;
    createDateFrom: Date;
    lineManagerId: string;
    projectId: string;
    boardTypeId: string;
    NoOfReplansMin: number;
    NoOfReplansMax: number;
    Showcomments: boolean;
    IsTableView: boolean;
    GoalSearchText: string;
    UserStorySearchText: string;
    BranchId: string;
    VerifiedBy: string;
    VerifiedOn: Date;
    IsGoalDealyed: boolean;
    IsUserStoryDealyed: boolean;
    UserStoryTypeId: string;
    UserStoryPriorityId: string;
    isSprintUserStory: boolean;
    isProject: boolean;
    columns: any;
    state: any;
    hiddenColumnList: any;
    excelColumnList: any;
    timeZone: number;
}